+++ returns/primitives/exceptions.py
@@ -2,18 +2,18 @@
 
 
 class UnwrapFailedError(Exception):
-    """Raised when a monad can not be unwrapped into a meaningful value."""
+    """Raised when a container can not be unwrapped into a meaningful value."""
 
-    def __init__(self, monad):
+    def __init__(self, container):
         """
-        Saves halted monad in the inner state.
+        Saves halted container in the inner state.
 
-        So, this monad can later be unpacked from this exception
+        So, this container can later be unpacked from this exception
         and used as a regular value.
         """
         super().__init__()
-        self.halted_monad = monad
+        self.halted_container = container
 
 
 class ImmutableStateError(Exception):
-    """Raised when a monad is forced to be mutated."""
+    """Raised when a container is forced to be mutated."""