+++ tests/test_functions/test_pipeline.py
@@ -21,15 +21,15 @@ def _transformation(number: int) -> Success[int]:
     return Success(-number)
 
 
-def test_do_notation_success():
-    """Ensures that do notation works well for Success."""
+def test_pipeline_success():
+    """Ensures that pipeline works well for Success."""
     assert _example1(5) == Success(6)
     assert _example1(1).unwrap() == 2
     assert _example1(9).bind(_transformation).value_or(None) == -10
 
 
-def test_do_notation_failure():
-    """Ensures that do notation works well for Failure."""
+def test_pipeline_failure():
+    """Ensures that pipeline works well for Failure."""
     assert _example1(0) == Failure('E')
     assert _example2(0) == Failure(0)
     assert _example2(1).rescue(_transformation).unwrap() == -1