+++ tests/test_primitives/test_monad.py
@@ -2,10 +2,10 @@
 
 import pytest
 
-from returns.primitives.monad import Monad
+from returns.primitives.container import Container
 
 
 def test_abstract_class():
-    """Ensures that Monad can not be instanciated."""
+    """Ensures that Container can not be instanciated."""
     with pytest.raises(TypeError):
-        Monad(1)  # type: ignore
+        Container(1)  # type: ignore